+++ systemtest/src/test/java/io/strimzi/systemtest/security/oauth/OauthAuthorizationST.java
@@ -84,7 +84,7 @@ public class OauthAuthorizationST extends OauthAbstractST {
 
         teamAOauthClientJob = new KafkaOauthClientsResource(teamAOauthClientJob, TOPIC_NAME, "a-consumer_group");
         teamAOauthClientJob.producerStrimziOauthTls(CLUSTER_NAME).done();
-        assertThrows(WaitException.class, () -> ClientUtils.waitForClientFailure(TEAM_A_PRODUCER_NAME, NAMESPACE, 30_000));
+        assertThrows(WaitException.class, () -> JobUtils.waitForJobFailure(TEAM_A_PRODUCER_NAME, NAMESPACE, 30_000));
         JobUtils.deleteJob(NAMESPACE, TEAM_A_PRODUCER_NAME);
 
         String topicXName = TOPIC_X + "-example-1";
@@ -92,7 +92,7 @@ public class OauthAuthorizationST extends OauthAbstractST {
 
         teamAOauthClientJob = new KafkaOauthClientsResource(teamAOauthClientJob, topicXName, "a-consumer_group");
         teamAOauthClientJob.producerStrimziOauthTls(CLUSTER_NAME).done();
-        assertThrows(WaitException.class, () -> ClientUtils.waitForClientFailure(TEAM_A_PRODUCER_NAME, NAMESPACE, 30_000));
+        assertThrows(WaitException.class, () -> JobUtils.waitForJobFailure(TEAM_A_PRODUCER_NAME, NAMESPACE, 30_000));
         JobUtils.deleteJob(NAMESPACE, TEAM_A_PRODUCER_NAME);
 
         // Team A can not create topic starting with 'x-' only write to existing on
@@ -119,7 +119,7 @@ public class OauthAuthorizationST extends OauthAbstractST {
         // team A client shouldn't be able to consume messages with wrong consumer group
         teamAOauthClientJob = new KafkaOauthClientsResource(teamAOauthClientJob, TOPIC_A, "bad_consumer_group");
         teamAOauthClientJob.consumerStrimziOauthTls(CLUSTER_NAME).done();
-        assertThrows(WaitException.class, () -> ClientUtils.waitForClientFailure(TEAM_A_CONSUMER_NAME, NAMESPACE, 30_000));
+        assertThrows(WaitException.class, () -> JobUtils.waitForJobFailure(TEAM_A_CONSUMER_NAME, NAMESPACE, 30_000));
         JobUtils.deleteJob(NAMESPACE, TEAM_A_PRODUCER_NAME);
 
         // team A client should be able to consume messages with correct consumer group
@@ -135,7 +135,7 @@ public class OauthAuthorizationST extends OauthAbstractST {
         LOGGER.info("Sending {} messages to broker with topic name {}", MESSAGE_COUNT, TOPIC_NAME);
         // Producer will not produce messages because authorization topic will failed. Team A can write only to topic starting with 'x-'
         teamBOauthClientJob.producerStrimziOauthTls(CLUSTER_NAME).done();
-        assertThrows(WaitException.class, () -> ClientUtils.waitForClientFailure(TEAM_B_PRODUCER_NAME, NAMESPACE, 30_000));
+        assertThrows(WaitException.class, () -> JobUtils.waitForJobFailure(TEAM_B_PRODUCER_NAME, NAMESPACE, 30_000));
         JobUtils.deleteJob(NAMESPACE, TEAM_B_PRODUCER_NAME);
 
         LOGGER.info("Sending {} messages to broker with topic name {}", MESSAGE_COUNT, TOPIC_B);
@@ -176,7 +176,7 @@ public class OauthAuthorizationST extends OauthAbstractST {
         teamBOauthClientJob = new KafkaOauthClientsResource(teamBOauthClientJob, USER_NAME);
 
         teamBOauthClientJob.producerStrimziOauthTls(CLUSTER_NAME).done();
-        assertThrows(WaitException.class, () -> ClientUtils.waitForClientFailure(TEAM_B_PRODUCER_NAME, NAMESPACE, 30_000));
+        assertThrows(WaitException.class, () -> JobUtils.waitForJobFailure(TEAM_B_PRODUCER_NAME, NAMESPACE, 30_000));
         JobUtils.deleteJob(NAMESPACE, TEAM_B_PRODUCER_NAME);
 
         LOGGER.info("Verifying that team A is not able read to topic starting with 'x-' because in kafka cluster" +
@@ -186,7 +186,7 @@ public class OauthAuthorizationST extends OauthAbstractST {
         teamAOauthClientJob = new KafkaOauthClientsResource(teamAOauthClientJob, USER_NAME);
 
         teamAOauthClientJob.consumerStrimziOauthTls(CLUSTER_NAME).done();
-        assertThrows(WaitException.class, () -> ClientUtils.waitForClientFailure(TEAM_A_CONSUMER_NAME, NAMESPACE, 30_000));
+        assertThrows(WaitException.class, () -> JobUtils.waitForJobFailure(TEAM_A_CONSUMER_NAME, NAMESPACE, 30_000));
         JobUtils.deleteJob(NAMESPACE, TEAM_A_CONSUMER_NAME);
 
         Map<String, String> kafkaPods = StatefulSetUtils.ssSnapshot(KafkaResources.kafkaStatefulSetName(CLUSTER_NAME));
